full functional without some logic such as payment calculations

+++ HealthCareMonitoring/Shared/Domain/Payment.cs
@@ -10,7 +10,7 @@ namespace HealthCareMonitoring.Shared.Domain
     {
         public virtual Prescription? Prescription { get; set; }
         public int? PrescriptionId { get; set; }
-        public int? AmtPaid { get; set; }
-        public int? AmtPayable { get; set; }
+        public double? AmtPaid { get; set; }
+        public double? AmtPayable { get; set; }
     }
 }
